+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * For CDK projects, check for the use of Yarn workspaces and compile all workspace packages before running CDK commands. This allows CDK code to depend on externalized local packages, such as a "constructs" module.
   * This feature can be disabled by setting a `COIN_SKIP_WORKSPACE_BUILD` environment variable to "y".
+* Added new make commands: `switch-current-environment-to` and `switch-current-environment-to-conditionally`
 
 # November 2025